pyproject.toml
[tool.poetry]
name = "velobus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
django-vitevue = "^0.7.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
ipython = "^8.4.0"

[tool.black]
line-length = 90

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
